// src/enums/chat.enum.ts
export enum ConversationType {
  DM = 'dm',
  GROUP = 'group',
}

export enum ConversationMemberRole {
  MEMBER = 'member',
  ADMIN = 'admin',
}

export enum MessageKind {
  TEXT = 'text',
  IMAGE = 'image',
  FILE = 'file',
  SYSTEM = 'system',
}

export enum MessageReceiptStatus {
  DELIVERED = 'delivered',
  READ = 'read',
}